backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.8.1
httpx==0.28.1
python-multipart==0.0.20
google-genai==1.31.0
